Extract MemoryCompressedSliceOutput
