+++ lib/trino-hive-formats/src/main/java/io/trino/hive/formats/compression/Compressor.java
@@ -13,97 +13,7 @@
  */
 package io.trino.hive.formats.compression;
 
-import io.airlift.slice.Slice;
-
-import java.io.IOException;
-import java.io.OutputStream;
-import java.util.List;
-import java.util.function.Supplier;
-
-import static com.google.common.base.Preconditions.checkState;
-import static java.util.Objects.requireNonNull;
-
 public interface Compressor
 {
-    CompressedSliceOutput createCompressedSliceOutput(int minChunkSize, int maxChunkSize);
-
-    // This specialized SliceOutput has direct access buffered output slices to
-    // report buffer sizes and to get he final output.  Additionally, a new
-    // CompressedSliceOutput can be created that reuses the underlying output
-    // buffer
-    final class CompressedSliceOutput
-            extends BufferedOutputStreamSliceOutput
-    {
-        private final ChunkedSliceOutput bufferedOutput;
-        private final Supplier<CompressedSliceOutput> resetFactory;
-        private final Runnable onDestroy;
-        private boolean closed;
-        private boolean destroyed;
-
-        /**
-         * @param compressionStream the compressed output stream to delegate to
-         * @param bufferedOutput the output for the compressionStream
-         * @param resetFactory the function to create a new CompressedSliceOutput that reuses the bufferedOutput
-         * @param onDestroy used to cleanup the compression when done
-         */
-        public CompressedSliceOutput(OutputStream compressionStream, ChunkedSliceOutput bufferedOutput, Supplier<CompressedSliceOutput> resetFactory, Runnable onDestroy)
-        {
-            super(compressionStream);
-            this.bufferedOutput = requireNonNull(bufferedOutput, "bufferedOutput is null");
-            this.resetFactory = requireNonNull(resetFactory, "resetFactory is null");
-            this.onDestroy = requireNonNull(onDestroy, "onDestroy is null");
-        }
-
-        @Override
-        public long getRetainedSize()
-        {
-            return super.getRetainedSize() + bufferedOutput.getRetainedSize();
-        }
-
-        public int getCompressedSize()
-        {
-            checkState(closed, "Stream has not been closed");
-            checkState(!destroyed, "Stream has been destroyed");
-            return bufferedOutput.size();
-        }
-
-        public List<Slice> getCompressedSlices()
-        {
-            checkState(closed, "Stream has not been closed");
-            checkState(!destroyed, "Stream has been destroyed");
-            return bufferedOutput.getSlices();
-        }
-
-        public CompressedSliceOutput createRecycledCompressedSliceOutput()
-        {
-            checkState(closed, "Stream has not been closed");
-            checkState(!destroyed, "Stream has been destroyed");
-            destroyed = true;
-            return resetFactory.get();
-        }
-
-        @Override
-        public void close()
-                throws IOException
-        {
-            if (!closed) {
-                closed = true;
-                super.close();
-            }
-        }
-
-        public void destroy()
-                throws IOException
-        {
-            if (!destroyed) {
-                destroyed = true;
-                try {
-                    close();
-                }
-                finally {
-                    onDestroy.run();
-                }
-            }
-        }
-    }
+    MemoryCompressedSliceOutput createCompressedSliceOutput(int minChunkSize, int maxChunkSize);
 }
